Rangement des classes dans grid.hpp / champ.hpp / config.hpp
